fixed bug when computing train_idcs and test_idcs for hyper param search

+++ evaluate.py
@@ -82,31 +82,23 @@ def setup_odeformer(trainer) -> SymbolicTransformerRegressor:
         params=trainer.params
     )
 
-
 def read_file(filename, label="target", sep=None):
-
     if filename.endswith("gz"):
         compression = "gzip"
     else:
         compression = None
-
     if sep:
         input_data = pd.read_csv(filename, sep=sep, compression=compression)
     else:
         input_data = pd.read_csv(
             filename, sep=sep, compression=compression, engine="python"
         )
-
     feature_names = [x for x in input_data.columns.values if x != label]
     feature_names = np.array(feature_names)
-
     X = input_data.drop(label, axis=1).values.astype(float)
     y = input_data[label].values
-
     assert X.shape[1] == feature_names.shape[0]
-
     return X, y, feature_names
-
 
 class Evaluator(object):
 
@@ -179,8 +171,15 @@ class Evaluator(object):
                                 "baseline_hyper_opt_eval_fraction is None. "\
                                 f"Using baseline_hyper_opt_eval_fraction = {baseline_hyper_opt_eval_fraction} instead."
                             )
-                        train_idcs = np.arange(int(np.floor(1-baseline_hyper_opt_eval_fraction*len(_times))))
-                        test_idcs = np.arange(int(np.floor(1-baseline_hyper_opt_eval_fraction*len(_times))), len(_times))
+                        train_idcs = np.arange(
+                            int(np.floor((1-baseline_hyper_opt_eval_fraction)*len(_times))), 
+                            dtype=int
+                        )
+                        test_idcs = np.arange(
+                            int(np.floor((1-baseline_hyper_opt_eval_fraction)*len(_times))), 
+                            len(_times),
+                            dtype=int
+                        )
                         _model = self.model.get_grid_search(train_idcs, test_idcs)
                         _model.fit(_times, _trajectory)
                         _all_candidates = _model.best_estimator_._get_equations()
